pyproject.toml
[tool.poetry]
name = "aniduel"
version = "0.1.0"
description = ""
authors = ["justanotherbyte <wishymovies@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
"discord.py" = "^2.0.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
